requirements.txt
flask>=2.3.0
python-dotenv>=1.0.0
flask-cors>=4.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
